LayoutTests/accessibility/mac/aria-liveregion-on-image-expected.txt


This tests that ARIA live regions that is on an image will work when the alt tag is changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
